requirements.txt
Django==1.5.1
Unidecode==0.04.12
django-email-extras==0.1.12
django-forms-builder==0.8.4
python-gnupg==0.3.3
sphinx-me==0.1.2
